+++ server/src/engineModel/EngineModel.ts
@@ -1,13 +1,12 @@
 import * as fs from 'fs';
-import { Parser } from 'xml2js';
+import { XmlDocument, XmlElement } from 'xmldoc';
 
 import { YmlCompletionItemsProvider } from '../completion/YmlCompletionItemsProvider';
-import { connection } from '../server';
+import { connection } from '../constants';
+import { YmlDefinitionProvider } from '../definitions/YmlDefinitionProvider';
 import { YmlAttribute, YmlClass, YmlFunction } from '../yml-objects';
 import { AbstractYmlFunction } from '../yml-objects/AbstractYmlFunction';
 import { YmlMethod } from '../yml-objects/YmlMethod';
-
-const parser = new Parser();
 
 /**
  * Parse `predefinedObjects.xml` and contain the classes and functions that can be found in it.
@@ -24,7 +23,11 @@ export class EngineModel {
     private classes: YmlClass[] = [];
     private functions: YmlFunction[] = [];
 
-    constructor(public uri: string, public completionProvider: YmlCompletionItemsProvider) {}
+    constructor(
+        public uri: string,
+        public completionProvider: YmlCompletionItemsProvider,
+        public definitionProvider: YmlDefinitionProvider,
+    ) {}
 
     /**
      * Parse the file that is at `uri`.
@@ -49,21 +52,22 @@ export class EngineModel {
         this.completionProvider = completionProvider;
         this.loadPredefinedObjects();
     }
-    private buildYmlFunction(func: any, sourceType?: string): AbstractYmlFunction {
+    private buildYmlFunction(func: XmlElement, sourceType?: string): AbstractYmlFunction {
         let ymlAbstractFunction: AbstractYmlFunction;
         const documentation: string = getDocValue(func);
         if (sourceType) {
-            ymlAbstractFunction = new YmlMethod(func.$.ident, this.uri);
+            ymlAbstractFunction = new YmlMethod(func.attr.ident, this.uri);
             this.functions.push(ymlAbstractFunction);
             ymlAbstractFunction.setUserInformations(
                 `(function) [${sourceType}].${ymlAbstractFunction.label}`,
                 documentation,
             );
         } else {
-            ymlAbstractFunction = new YmlFunction(func.$.ident, this.uri);
+            ymlAbstractFunction = new YmlFunction(func.attr.ident, this.uri);
             ymlAbstractFunction.setUserInformations(`(function) [static] ${ymlAbstractFunction.label}`, documentation);
         }
         ymlAbstractFunction.data = `id_${sourceType ?? 'static'}_${ymlAbstractFunction.label}`;
+        ymlAbstractFunction.setDefinitionLocationFromLocation(this.buildLocationFromXmlElement(func));
         return ymlAbstractFunction;
     }
 
@@ -72,46 +76,49 @@ export class EngineModel {
      *
      * @param yclass The data structure resulting from the parsing of `predefinedObjects.xml`.
      */
-    private buildYmlClass(yclass: any): void {
-        const ymlClass: YmlClass = new YmlClass(yclass.$.ident, this.uri);
-        if (yclass.extends) {
-            yclass.extends.forEach((ext) => {
-                ymlClass.extends.push(ext);
+    private buildYmlClass(yclass: XmlElement): void {
+        const ymlClass: YmlClass = new YmlClass(yclass.attr.ident, this.uri);
+        yclass.childrenNamed('extends')?.forEach((ext) => {
+            ext.children.forEach((element) => {
+                ymlClass.extends.push(element.toString());
             });
-        }
-        if (yclass.attribute) {
-            yclass.attribute.forEach((att) => {
-                const attribute = this.buildAttribute(att, ymlClass.label);
-                ymlClass.attributes.push(attribute);
-                this.completionProvider.addCompletionItem(attribute);
-            });
-        }
-        if (yclass.method) {
-            yclass.method.forEach((meth) => {
-                const method = this.buildYmlFunction(meth, ymlClass.label);
-                ymlClass.methods.push(method);
-                this.completionProvider.addCompletionItem(method);
-            });
-        }
+        });
+        yclass.childrenNamed('attribute').forEach((att) => {
+            const attribute = this.buildAttribute(att, ymlClass.label);
+            attribute.setDefinitionLocationFromLocation(this.buildLocationFromXmlElement(att));
+            ymlClass.attributes.push(attribute);
+            this.completionProvider.addCompletionItem(attribute);
+            this.definitionProvider.addDefinition(attribute);
+        });
+        yclass.childrenNamed('method').forEach((meth) => {
+            const method = this.buildYmlFunction(meth, ymlClass.label);
+            method.setDefinitionLocationFromLocation(this.buildLocationFromXmlElement(meth));
+            ymlClass.methods.push(method);
+            this.completionProvider.addCompletionItem(method);
+            this.definitionProvider.addDefinition(method);
+        });
         const documentation: string = getDocValue(yclass);
         ymlClass.setUserInformations(`(class) ${ymlClass.label}`, documentation);
+        ymlClass.setDefinitionLocationFromLocation(this.buildLocationFromXmlElement(yclass));
         this.classes.push(ymlClass);
         this.enrichYmlClass(ymlClass);
         this.completionProvider.addCompletionItem(ymlClass);
+        this.definitionProvider.addDefinition(ymlClass);
     }
 
-    private buildAttribute(attributeXmlElement: any, sourceType: string): YmlAttribute {
-        const attribute = new YmlAttribute(attributeXmlElement.$.ident, this.uri);
+    private buildAttribute(attributeXmlElement: XmlElement, sourceType: string): YmlAttribute {
+        const attribute = new YmlAttribute(attributeXmlElement.attr.ident, this.uri);
         const documentation: string = getDocValue(attributeXmlElement);
         attribute.setUserInformations(`(property) [${sourceType}].${attribute.label}`, documentation);
-        attributeXmlElement.return.forEach((returnType) => {
-            if (returnType.domains) {
-                attribute.domains = returnType.domains;
-            }
-            if (returnType.domainsLevel2) {
-                attribute.domainsLevel2 = returnType.domainsLevel2;
-            }
-        });
+        const returnTypes = attributeXmlElement.childNamed('return');
+        const domains = returnTypes.childrenNamed('domains');
+        if (!!domains) {
+            attribute.domains = domains.map((_currentDomain) => _currentDomain.val);
+        }
+        const domainsLevel2 = returnTypes.childrenNamed('domainsLevel2');
+        if (!!domainsLevel2) {
+            attribute.domainsLevel2 = domainsLevel2.map((_currentDomain) => _currentDomain.val);
+        }
         return attribute;
     }
 
@@ -123,25 +130,23 @@ export class EngineModel {
     }
 
     public parsePredefinedObjectsFileContent(err, data: Buffer | string): void {
-        parser.parseString(data, (parseErr, predefinedObjects) => {
-            if (err != null) {
-                connection.console.error(`Something went wrong during YE model import:\n ${parseErr}`);
-            } else if (predefinedObjects == null) {
-                connection.console.error('Something went wrong during YE model import. Your file seems empty.');
-            } else {
-                try {
-                    const dataAndFeatures = predefinedObjects['data-and-features'];
-                    this.importClasses(dataAndFeatures);
-                    this.importFunctions(dataAndFeatures);
-                    this.importTags(dataAndFeatures);
-                } catch (importErr) {
-                    connection.console.error(`Something went wrong during YE model import:\n ${importErr}`);
-                }
+        if (!!err) {
+            connection.console.error(`Something went wrong during YE model import:\n ${err}`);
+        } else if (data === null || data.length === 0) {
+            connection.console.error('Something went wrong during YE model import. Your file seems empty.');
+        } else {
+            try {
+                const document: XmlDocument = new XmlDocument(data.toString());
+                this.importClasses(document);
+                this.importFunctions(document);
+                this.importTags(document);
+            } catch (importErr) {
+                connection.console.error(`Something went wrong during YE model import:\n ${importErr}`);
             }
-            connection.console.log(
-                `Done with classes size=${this.classes.length}\nfunctions size=${this.functions.length}`,
-            );
-        });
+        }
+        connection.console.log(
+            `Done with classes size=${this.classes.length}\nfunctions size=${this.functions.length}`,
+        );
     }
 
     /**
@@ -168,16 +173,13 @@ export class EngineModel {
      *
      * @param dataAndFeatures The data structure resulting from the parsing of `predefinedObjects.xml`.
      */
-    private importClasses(dataAndFeatures: any): void {
+    private importClasses(dataAndFeatures: XmlDocument): void {
         connection.console.log('Importing classes from Yseop Engine model.');
         // Get the single “classes” element.
-        const classesByPackage = dataAndFeatures.classes[0].package;
+        const classesByPackage: XmlElement[] = dataAndFeatures.childNamed('classes').childrenNamed('package');
         classesByPackage.forEach((ypackage) => {
-            if (ypackage.class == null) {
-                return;
-            }
             // empty package
-            ypackage.class.forEach((yclass) => {
+            ypackage.childrenNamed('class').forEach((yclass) => {
                 this.buildYmlClass(yclass);
             });
         });
@@ -207,12 +209,13 @@ export class EngineModel {
      *
      * @param dataAndFeatures The data structure resulting from the parsing of `predefinedObjects.xml`.
      */
-    private importFunctions(dataAndFeatures: any): any {
+    private importFunctions(dataAndFeatures: XmlElement): any {
         // Get the single “functions” element.
-        const ymlFunctions = dataAndFeatures.functions[0];
-        ymlFunctions.function.forEach((func) => {
+        const ymlFunctions = dataAndFeatures.childNamed('functions');
+        ymlFunctions.childrenNamed('function').forEach((func) => {
             const yFunction = this.buildYmlFunction(func);
             this.completionProvider.addCompletionItem(yFunction);
+            this.definitionProvider.addDefinition(yFunction);
         });
     }
 
@@ -237,16 +240,17 @@ export class EngineModel {
      *
      * @param dataAndFeatures The data structure resulting from the parsing of `predefinedObjects.xml`.
      */
-    private importTags(dataAndFeatures: any): void {
+    private importTags(dataAndFeatures: XmlElement): void {
         // Get the single “text-tags” element.
-        const textTags = dataAndFeatures['text-tags'][0];
-        textTags.tag.forEach((_tag) => {
-            if (!_tag.$.ident) {
+        const textTags = dataAndFeatures.childNamed('text-tags');
+        textTags.childrenNamed('tag').forEach((_tag) => {
+            if (!_tag.attr.ident) {
                 // invalid tag element without identifier
                 return;
             }
             const textTag = this.buildYmlFunction(_tag);
             this.completionProvider.addCompletionItem(textTag);
+            this.definitionProvider.addDefinition(textTag);
         });
     }
 
@@ -262,25 +266,45 @@ export class EngineModel {
             yclass.detail = yclass.detail.concat(` extends ${yclass.extends}`);
         }
     }
+
+    private buildLocationFromXmlElement(xmlElement: XmlElement) {
+        return {
+            range: {
+                end: {
+                    character: xmlElement.column,
+                    line: xmlElement.line,
+                },
+                start: {
+                    character: xmlElement.column,
+                    line: xmlElement.line,
+                },
+            },
+            uri: this.uri,
+        };
+    }
 }
 
 /**
- * Returns the value of the first child `doc` element within a given XML element, if any.
+ * Returns the value of the cdata contained in the first child `doc` element within a given XML element, if any.
  *
- * We expect it to be given as the following JSON representation, obtained thanks to `xml2js`.
+ * We expect it to be given as the following representation.
  *
- * ```JSON
- * {
- *     doc: ["documentation"]
- * }
+ * ```XML
+ * <doc><![CDATA[documentation]]></doc>
  * ```
  *
- * Using this function on the JSON representation above will return `"documentation"`.
+ * Using this function on the XML representation above will return `"documentation"`.
+ * Any other representation of the node will result in a `null` value.
  *
- * @param xmlElement a JSON representation of an XML tag.
+ * @param xmlElement a representation of an XML tag.
  *
  * @return the xmlElement's documentation or `null`.
  */
-function getDocValue(xmlElement: any): string {
-    return !!xmlElement.doc ? xmlElement.doc[0] : null;
+function getDocValue(xmlElement: XmlElement): string {
+    const doc = xmlElement.childNamed('doc');
+    const cdata = !!doc ? doc.firstChild : null;
+    if (!cdata || cdata.type !== 'cdata') {
+        return null;
+    }
+    return cdata.cdata;
 }