+++ server/src/visitors/YmlFunctionVisitor.ts
@@ -1,5 +1,6 @@
 import { ParserRuleContext } from 'antlr4ts';
-import { TextDocument, TextEdit } from 'vscode-languageserver';
+import { Diagnostic, DiagnosticSeverity, TextEdit } from 'vscode-languageserver';
+import { TextDocument } from 'vscode-languageserver-textdocument';
 
 import { YmlCompletionItemsProvider } from '../completion/YmlCompletionItemsProvider';
 import { YmlDefinitionProvider } from '../definitions';
@@ -11,6 +12,7 @@ import {
     ElseExpressionContext,
     ElseIfExpressionContext,
     FunctionContext,
+    GranuleContext,
     IfExprBlockContext,
     Instruction_breakContext,
     Instruction_forallContext,
@@ -26,9 +28,11 @@ import {
     MandatoryArgDeclContext,
     MemberDeclarationContext,
     VariableBlockContentContext,
+    YmlIdContext,
 } from '../grammar';
 import { IDocumentFormatSettings } from '../settings/Settings';
 import { AbstractYmlFunction, YmlArgument, YmlFunction, YmlObjectInstance } from '../yml-objects';
+import { AbstractYmlObject } from '../yml-objects/AbstractYmlObject';
 import { YmlBaseVisitor } from './YmlBaseVisitor';
 import { getDocumentation, getType } from './YmlVisitorHelper';
 
@@ -48,6 +52,8 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
 
     private func: AbstractYmlFunction;
 
+    public unusedVariables: AbstractYmlObject[];
+
     /**
      * State used when computing the cognitive complexity.
      * The complexity should increase when the operator changes
@@ -62,8 +68,10 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
         public filePossibleEdits: TextEdit[] = [],
         public document: TextDocument = null,
         public docFormatSettings: IDocumentFormatSettings = null,
+        public diagnostics?: Diagnostic[],
     ) {
         super(completionProvider, uri);
+        this.unusedVariables = [];
     }
     public visitFunction(node: FunctionContext): void {
         this.scopeStartOffset = 0;
@@ -105,6 +113,15 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
          * Look for the function's arguments and local variables.
          */
         this.visitChildren(node);
+        for (const unusedVariable of this.unusedVariables) {
+            this.diagnostics?.push(
+                Diagnostic.create(
+                    unusedVariable.definitionLocation.range,
+                    `The variable “${unusedVariable.label}” is probably unused.`,
+                    DiagnosticSeverity.Information,
+                ),
+            );
+        }
     }
 
     /**
@@ -115,6 +132,20 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
         const arg = new YmlArgument(node._argName.text, this.uri);
         arg.enrichWith(null, node._argType.text, this.func, this.scopeStartOffset, this.scopeEndOffset);
         this.completionProvider.addCompletionItem(arg);
+        this.unusedVariables.push(arg);
+        arg.definitionLocation = {
+            range: {
+                end: {
+                    character: node.stop.charPositionInLine + node.stop.text.length,
+                    line: node.stop.line - 1,
+                },
+                start: {
+                    character: node.start.charPositionInLine,
+                    line: node.start.line - 1,
+                },
+            },
+            uri: this.uri,
+        };
     }
     /**
      * Visit the member declarations that are in `local` and `args` blocks of functions.
@@ -144,7 +175,21 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
         const doc = getDocumentation(node.field());
         const type = getType(node.field(), node._type.text);
         variable.enrichWith(doc, type, this.func, this.scopeStartOffset, this.scopeEndOffset);
+        variable.definitionLocation = {
+            range: {
+                end: {
+                    character: node.stop.charPositionInLine + node.stop.text.length,
+                    line: node.stop.line - 1,
+                },
+                start: {
+                    character: node.start.charPositionInLine,
+                    line: node.start.line - 1,
+                },
+            },
+            uri: this.uri,
+        };
         this.completionProvider.addCompletionItem(variable);
+        this.unusedVariables.push(variable);
     }
 
     /**
@@ -157,6 +202,14 @@ export class YmlFunctionVisitor extends YmlBaseVisitor {
     private isMethodInstanciation(fullName: string): boolean {
         // Check that there is a “::”, with at least one character before it.
         return fullName.includes('::', 1);
+    }
+
+    public visitYmlId(node: YmlIdContext) {
+        this.unusedVariables = this.unusedVariables.filter((elem) => elem.label !== node.text);
+    }
+
+    public visitGranule(node: GranuleContext): void {
+        this.unusedVariables = this.unusedVariables.filter((elem) => !node.text.includes(elem.label));
     }
 
     public visitInstruction_try_catch(node: Instruction_try_catchContext) {